+++ src/backend/marsha/core/tests/test_models_passport.py
@@ -4,7 +4,12 @@ from django.test import TestCase
 
 from safedelete.models import HARD_DELETE
 
-from ..factories import ConsumerSiteLTIPassportFactory, PlaylistLTIPassportFactory
+from ..factories import (
+    ConsumerSiteFactory,
+    ConsumerSiteLTIPassportFactory,
+    PlaylistFactory,
+    PlaylistLTIPassportFactory,
+)
 from ..models import LTIPassport
 
 
@@ -26,27 +31,58 @@ class LTIPassportModelsTestCase(TestCase):
         pl_passport.delete()
         self.assertEqual(str(pl_passport), "clé_pl [pl][deleted]")
 
-    def test_models_lti_passport_fields_nothing_required(self):
-        """The LTIPassport model can be instantiated without setting any field.
+    def test_models_lti_passport_fields_playlist_or_consumer_site_required(self):
+        """Creating an LTIPassport requires either a consumer site or a playlist."""
+        with self.assertRaises(ValidationError) as context:
+            LTIPassport.objects.create()
+        self.assertEqual(
+            context.exception.messages,
+            ["You must set either a Consumer Site or a Playlist."],
+        )
 
-        In particular:
-        - the "consumer_site" and "playlist" foreign keys are not required,
-        - the "oauth_consumer_key" and "shared_secret" fields are auto-generated,
-        - the "is_enabled" flag defaults to True.
+    def test_models_lti_passport_fields_playlist(self):
+        """It should be possible to create an LTIPassport with just a playlist.
+
+        The "oauth_consumer_key" and "shared_secret" fields are auto-generated,
+        The "is_enabled" flag defaults to True.
         """
-        lti_passport = LTIPassport.objects.create()
+        playlist = PlaylistFactory()
+        lti_passport = LTIPassport.objects.create(playlist=playlist)
         self.assertIsNone(lti_passport.consumer_site)
+        self.assertEqual(len(lti_passport.oauth_consumer_key), 20)
+        self.assertEqual(len(lti_passport.shared_secret), 40)
+        self.assertTrue(lti_passport.is_enabled)
+
+    def test_models_lti_passport_fields_consumer_site(self):
+        """It should be possible to create an LTIPassport with just a consumer site.
+
+        The "oauth_consumer_key" and "shared_secret" fields are auto-generated,
+        The "is_enabled" flag defaults to True.
+        """
+        consumer_site = ConsumerSiteFactory()
+        lti_passport = LTIPassport.objects.create(consumer_site=consumer_site)
         self.assertIsNone(lti_passport.playlist)
         self.assertEqual(len(lti_passport.oauth_consumer_key), 20)
         self.assertEqual(len(lti_passport.shared_secret), 40)
         self.assertTrue(lti_passport.is_enabled)
 
+    def test_models_lti_passport_fields_playlist_and_consumer_site_both(self):
+        """Creating an LTIPassport with both playlist and consumer site should raise an error."""
+        consumer_site = ConsumerSiteFactory()
+        playlist = PlaylistFactory()
+        with self.assertRaises(ValidationError) as context:
+            LTIPassport.objects.create(consumer_site=consumer_site, playlist=playlist)
+        self.assertEqual(
+            context.exception.messages,
+            ["You should set either a Consumer Site or a Playlist, but not both."],
+        )
+
     def test_models_lti_passport_fields_oauth_consumer_key_unique(self):
         """The "oauth_consumer_key" field should be unique."""
-        lti_passport = LTIPassport.objects.create()
+        lti_passport = ConsumerSiteLTIPassportFactory()
 
         with self.assertRaises(ValidationError) as context:
-            LTIPassport.objects.create(
+            ConsumerSiteLTIPassportFactory(
                 oauth_consumer_key=lti_passport.oauth_consumer_key
             )
         self.assertEqual(
@@ -56,7 +92,7 @@ class LTIPassportModelsTestCase(TestCase):
 
     def test_models_lti_passport_fields_oauth_consumer_key_unique_deleted(self):
         """The "oauth_consumer_key" field should even be unique including the deleted instances."""
-        lti_passport = LTIPassport.objects.create()
+        lti_passport = ConsumerSiteLTIPassportFactory()
         oauth_consumer_key = lti_passport.oauth_consumer_key
         lti_passport.delete()
         self.assertIsNotNone(
@@ -64,7 +100,7 @@ class LTIPassportModelsTestCase(TestCase):
         )
 
         with self.assertRaises(ValidationError) as context:
-            LTIPassport.objects.create(oauth_consumer_key=oauth_consumer_key)
+            ConsumerSiteLTIPassportFactory(oauth_consumer_key=oauth_consumer_key)
         self.assertEqual(
             context.exception.messages,
             ["LTI passport with this Oauth consumer key already exists."],